# eventops/backend/app/api/knowledge.py
import re
from pathlib import Path
from uuid import uuid4
from typing import Any

from fastapi import APIRouter, Depends, File, Form, HTTPException, UploadFile, status
from sqlalchemy import and_, false, or_, select, true
from sqlalchemy.ext.asyncio import AsyncSession

from .. import schemas
from ..auth import get_current_staff
from ..database import get_db
from ..models import ConfidentialityRule, KnowledgeBaseLink, Staff, Visibility
from ..rag import index_document_chunks, search_document_chunks
from .common import can_see_confidential, ensure_event_access

router = APIRouter(prefix="/events/{event_id}", tags=["knowledge"])


def _document_storage_dir() -> Path:
    return Path("storage/documents")


def _safe_filename(value: str) -> str:
    name = Path(value or "document").name
    return re.sub(r"[^a-zA-Z0-9_.-]", "_", name) or "document"


def _parse_tags(value: str | None) -> list[str]:
    if not value:
        return []
    return [item.strip() for item in value.split(",") if item.strip()]


async def _save_upload(file: UploadFile, destination: Path) -> bytes:
    content = await file.read()
    destination.parent.mkdir(parents=True, exist_ok=True)
    destination.write_bytes(content)
    return content


async def _kb_visibility_filter(db: AsyncSession, staff: Staff) -> Any:
    if staff.is_admin:
        return true()

    confidential_allowed = await can_see_confidential(db, staff)
    return or_(
        KnowledgeBaseLink.visibility == Visibility.public,
        and_(KnowledgeBaseLink.visibility == Visibility.role_only, true() if staff.role_id else false()),
        and_(KnowledgeBaseLink.visibility == Visibility.confidential, true() if confidential_allowed else false()),
    )


async def _require_admin(staff: Staff) -> None:
    if not staff.is_admin:
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Only admin can modify knowledge settings")


@router.get("/knowledge", response_model=list[schemas.KnowledgeBaseLink])
async def list_knowledge_links(
    event_id: int,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> list[KnowledgeBaseLink]:
    await ensure_event_access(event_id, current_staff)
    visible_clause = await _kb_visibility_filter(db, current_staff)
    result = await db.execute(
        select(KnowledgeBaseLink)
        .where(KnowledgeBaseLink.event_id == event_id, KnowledgeBaseLink.is_active.is_(True))
        .where(visible_clause)
        .order_by(KnowledgeBaseLink.id.asc())
    )
    return list(result.scalars().all())


@router.post("/knowledge", response_model=schemas.KnowledgeBaseLink, status_code=status.HTTP_201_CREATED)
async def create_knowledge_link(
    event_id: int,
    payload: schemas.KnowledgeBaseLinkCreate,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> KnowledgeBaseLink:
    await ensure_event_access(event_id, current_staff)
    await _require_admin(current_staff)
    link = KnowledgeBaseLink(event_id=event_id, **payload.model_dump())
    db.add(link)
    await db.commit()
    await db.refresh(link)
    return link


@router.get("/knowledge/search", response_model=list[schemas.DocumentChunkSearchResult])
async def search_knowledge(
    event_id: int,
    q: str,
    limit: int = 5,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> list[dict[str, Any]]:
    await ensure_event_access(event_id, current_staff)
    return await search_document_chunks(db, event_id=event_id, query=q, limit=max(1, min(limit, 20)))


@router.post("/knowledge/upload", response_model=schemas.KnowledgeBaseLink, status_code=status.HTTP_201_CREATED)
async def upload_knowledge_document(
    event_id: int,
    file: UploadFile = File(...),
    title: str | None = Form(default=None),
    description: str | None = Form(default=None),
    tags: str | None = Form(default=None, description="Comma-separated tags"),
    visibility: Visibility = Form(default=Visibility.public),
    is_active: bool = Form(default=True),
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> KnowledgeBaseLink:
    await ensure_event_access(event_id, current_staff)
    await _require_admin(current_staff)

    document_id = uuid4().hex
    filename = _safe_filename(file.filename or "document")
    destination = _document_storage_dir() / str(event_id) / "knowledge" / f"{document_id}_{filename}"
    content = await _save_upload(file, destination)
    link = KnowledgeBaseLink(
        event_id=event_id,
        title=title or filename,
        url=str(destination),
        description=description,
        tags=_parse_tags(tags),
        is_active=is_active,
        visibility=visibility,
    )
    db.add(link)
    await db.flush()
    await index_document_chunks(
        db,
        event_id=event_id,
        content=content,
        source_title=link.title,
        source_url=link.url,
        knowledge_base_link_id=link.id,
        metadata={"filename": filename, "content_type": file.content_type},
    )
    await db.commit()
    await db.refresh(link)
    return link


@router.patch("/knowledge/{link_id}", response_model=schemas.KnowledgeBaseLink)
async def update_knowledge_link(
    event_id: int,
    link_id: int,
    payload: schemas.KnowledgeBaseLinkUpdate,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> KnowledgeBaseLink:
    await ensure_event_access(event_id, current_staff)
    await _require_admin(current_staff)
    link = await db.get(KnowledgeBaseLink, link_id)
    if link is None or link.event_id != event_id:
        raise HTTPException(status_code=404, detail="Knowledge link not found")

    for key, value in payload.model_dump(exclude_unset=True).items():
        setattr(link, key, value)
    await db.commit()
    await db.refresh(link)
    return link


@router.get("/confidentiality-rules", response_model=list[schemas.ConfidentialityRule])
async def list_confidentiality_rules(
    event_id: int,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> list[ConfidentialityRule]:
    await ensure_event_access(event_id, current_staff)
    if not await can_see_confidential(db, current_staff):
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No access to confidentiality rules")

    result = await db.execute(
        select(ConfidentialityRule)
        .where(ConfidentialityRule.event_id == event_id, ConfidentialityRule.is_active.is_(True))
        .order_by(ConfidentialityRule.severity.desc(), ConfidentialityRule.id.asc())
    )
    return list(result.scalars().all())


@router.post("/confidentiality-rules", response_model=schemas.ConfidentialityRule, status_code=status.HTTP_201_CREATED)
async def create_confidentiality_rule(
    event_id: int,
    payload: schemas.ConfidentialityRuleCreate,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> ConfidentialityRule:
    await ensure_event_access(event_id, current_staff)
    await _require_admin(current_staff)
    rule = ConfidentialityRule(event_id=event_id, **payload.model_dump())
    db.add(rule)
    await db.commit()
    await db.refresh(rule)
    return rule


@router.patch("/confidentiality-rules/{rule_id}", response_model=schemas.ConfidentialityRule)
async def update_confidentiality_rule(
    event_id: int,
    rule_id: int,
    payload: schemas.ConfidentialityRuleUpdate,
    db: AsyncSession = Depends(get_db),
    current_staff: Staff = Depends(get_current_staff),
) -> ConfidentialityRule:
    await ensure_event_access(event_id, current_staff)
    await _require_admin(current_staff)
    rule = await db.get(ConfidentialityRule, rule_id)
    if rule is None or rule.event_id != event_id:
        raise HTTPException(status_code=404, detail="Confidentiality rule not found")

    for key, value in payload.model_dump(exclude_unset=True).items():
        setattr(rule, key, value)
    await db.commit()
    await db.refresh(rule)
    return rule