# openapi.yaml
openapi: 3.1.0
info:
  title: EventOps AI
  description: |
    AI-штаб для координации команды на мероприятии.

    **Аутентификация:** все эндпоинты кроме `/auth/login` требуют
    заголовок `Authorization: Bearer <token>`.

    **Visibility-фильтрация:** бэкенд автоматически фильтрует тикеты
    и сообщения по правам текущего сотрудника. Фронт ничего
    дополнительно не фильтрует.
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Local dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Enums ────────────────────────────────────────────────────────────────
  schemas:
...
    Visibility:
      type: string
      enum: [public, role_only, confidential]
      description: |
        - `public`       — видят все участники мероприятия
        - `role_only`    — видит только своя роль + admin
        - `confidential` — только admin + роли с can_see_confidential
...
    EventCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: ICPC 2025 Semifinal
        description:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
          nullable: true
        end_time:
          type: string
          format: date-time
          nullable: true
    Event:
      allOf:
        - $ref: '#/components/schemas/EventCreate'
        - type: object
          required: [id, created_at]
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
...
    KnowledgeBaseLinkCreate:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
        url:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
          default: true
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: public
    KnowledgeBaseLinkUpdate:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        visibility:
          $ref: '#/components/schemas/Visibility'
    KnowledgeBaseLink:
      allOf:
        - $ref: '#/components/schemas/KnowledgeBaseLinkCreate'
        - type: object
          required: [id, event_id, created_at, updated_at]
          properties:
            id:
              type: integer
            event_id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    DocumentChunkSearchResult:
      type: object
      required: [id, event_id, content, source_title, chunk_index]
      properties:
        id:
          type: integer
        event_id:
          type: integer
        knowledge_base_link_id:
          type: integer
          nullable: true
        ticket_id:
          type: integer
          nullable: true
        content:
          type: string
        source_title:
          type: string
        source_url:
          type: string
          nullable: true
        chunk_index:
          type: integer
        score:
          type: number
          nullable: true

    # ─── Agent ──────────────────────────────────────────────────────────────
...
    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: Not found

# ─── Paths ───────────────────────────────────────────────────────────────────
paths:
...
  /events/{event_id}:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [events]
      summary: Мероприятие по id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
...
  /events/{event_id}/knowledge:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [knowledge]
      summary: Список материалов базы знаний
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeBaseLink'
    post:
      tags: [knowledge]
      summary: Добавить ссылку в базу знаний
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeBaseLinkCreate'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseLink'
  /events/{event_id}/knowledge/upload:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags: [knowledge]
      summary: Загрузить документ в базу знаний
      description: Multipart upload. Доступно только организатору/admin.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
                tags:
                  type: string
                  nullable: true
                  description: Comma-separated tags
                visibility:
                  $ref: '#/components/schemas/Visibility'
                is_active:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Документ загружен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseLink'
  /events/{event_id}/knowledge/search:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [knowledge]
      summary: Поиск по базе знаний через RAG
      description: В PostgreSQL использует pgvector + Yandex text-search-query embeddings; в SQLite падает обратно на текстовый поиск.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentChunkSearchResult'
  /events/{event_id}/knowledge/{link_id}:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: link_id
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [knowledge]
      summary: Обновить материал базы знаний
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeBaseLinkUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseLink'

  # ── Messages ──────────────────────────────────────────────────────────────
...